requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.8
pydantic==2.5.0
asyncio-throttle==1.0.2
python-dotenv==1.0.0
httpx==0.25.2